Moved convert_json + tests to lua and renamed json_to_lua
